add flexibility to screen res of KC_SCREEN

+++ components/gnuboy-esp32/lcd.h
@@ -1,3 +1,3 @@
 
-uint32_t *vidGetOverlayBuf();
+uint16_t *vidGetOverlayBuf();
 void vidRenderOverlay();
